pyproject.toml
[tool.poetry]
name = "aivillage1"
version = "0.1.0"
description = ""
authors = ["6-28-496 <gmaccaba@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
djangorestframework-simplejwt = "^5.3.1"
openai = "^1.52.0"
psycopg2-binary = "^2.9.10"
django-environ = "^0.11.2"
django-cors-headers = "^4.5.0"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
